datetime_helpers: Show end date for multi-day events with start time
In `datetime_helpers.rb`, improve formatting for events with start date, start time, and end date but no end time (Case 4).

Previously, this case only displayed the start date and time, even for multi-day events. This change adds a check to see if the start and end dates differ. If they do, the end date is now appended to the output string to indicate the duration of the multi-day event.